fixes minor issues in custom kernel stat

+++ test-zonal.py
@@ -62,7 +62,7 @@ if __name__ == '__main__':
     assert(H/zH == H//zH)
 
     # Values raster
-    values = xr.DataArray(np.arange(H * W).reshape(H, W))
+    values = xr.DataArray(np.arange(H * W, dtype=float).reshape(H, W))
     values = create_arr(values, backend=args.backend)
     # Zones raster
     zones = xr.DataArray(np.zeros(H * W).reshape(H, W))
